refactor: financial statement sql structure

+++ sql/04-financialstatement.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS financialstatements (
     id INT PRIMARY KEY generated always as identity,
     ticker_id INT REFERENCES tickers(id),
     statement_date DATE NOT NULL,
-    stocks_amount INT NOT NULL,
+    stocks_amount BIGINT NOT NULL,
     div_yield NUMERIC(5,2) NOT NULL,
     assets BIGINT NOT NULL,
     disponibility BIGINT NOT NULL,
